Add a 2nd serial feed to list to rx
Jarvis desk has two serial transmitters, one on the
controller and one on the handset. To listen to both
of them, we need two serial interfaces. Fortunately
the esp8266 has a nice interrupt-driven serial library
that lets us eavesdrop easily.

Refactor the serial buffer code so we can reuse it on
two different interfaces. Add a listener for the handset
interface's Tx pin to sniff those messages, too.