Add a "stop-moving" step when preset received
If we're traveling to some new place and a new preset command is
received, we wait until the traveling stops before commanding the
new preset. Add a step to try to stop the traveling motion actively
by pressing the "Memory" button and interrupting the desk's plans.

Also, shorten our restart delay to 3/4 second instead of 2 seconds.