Refactor serial protocol into separate files
Extract the protocol code out of the JarvisDesk struct so it can be more easily
managed, extended, and replaced.  In particular this should make it easier
for Karuption's <pederson26@gmail.com> change to be incorporated into
the project as it aims to work with the other protocol seen from time to time
(and all the time on his desk).