Remove Handset/Desk column from protocol tables
Tables are separated so all commands for handset are in one and all responses
from the desk are in another. There's no need to identify them all as H in one
and all as D in the other.

Also, realign the vertical bars in the md code for easier readability.